+++ .browserslistrc
@@ -9,6 +9,7 @@
 #   npx browserslist
 
 last 1 Chrome version
+last 1 Firefox version
 last 2 Edge major versions
 last 2 Safari major versions
 last 2 iOS major versions
