+++ angular.json
@@ -13,6 +13,7 @@
         "build": {
           "builder": "@angular-devkit/build-angular:browser",
           "options": {
+            
             "outputPath": "dist/e-kart",
             "index": "src/index.html",
             "main": "src/main.ts",
