+++ src/app/app.module.ts
@@ -5,20 +5,20 @@ import { BrowserModule } from '@angular/platform-browser';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import { DefaultComponent } from './components/default/default.component';
-import { DashboardComponent } from './components/dashboard/dashboard.component';
-import { PostsComponent } from './components/posts/posts.component';
-import { TopBarComponent } from './components/top-bar/top-bar.component';
-import { FooterComponent } from './components/footer/footer.component';
+import { DefaultComponent } from './components/admin-desk/default/default.component'
+import { DashboardComponent } from './components/admin-desk/dashboard/dashboard.component';
+import { PostsComponent } from './components/admin-desk/posts/posts.component';
+import { TopBarComponent } from './components/admin-desk/top-bar/top-bar.component';
+import { FooterComponent } from './components/admin-desk/footer/footer.component';
 import { MaterialModule } from './material/material/material.module';
-import { SidenavComponent } from './components/sidenav/sidenav.component';
+import { SidenavComponent } from './components/admin-desk/sidenav/sidenav.component';
 import { HttpClientModule } from '@angular/common/http';
-import { AddProductComponent } from './components/add-product/add-product.component';
+import { AddProductComponent } from './components/admin-desk/add-product/add-product.component';
 
 import { MatTableModule } from '@angular/material/table';
 import { MatPaginatorModule } from '@angular/material/paginator';
 import { MatSortModule } from '@angular/material/sort';
-import { AdminDashboardComponent } from './components/admin-dashboard/admin-dashboard.component';
+import { AdminDashboardComponent } from './components/admin-desk/admin-dashboard/admin-dashboard.component';
 import { AreaWidgetsComponent } from './components/widgets/area-widgets/area-widgets.component';
 import { CardWidgetComponent } from './components/widgets/card-widget/card-widget.component';
 import { PieWidgetComponent } from './components/widgets/pie-widget/pie-widget.component';
